retain configmap name for use in updating status as it is not always the same as address
